+++ src/symbolize/symbolizer.rs
@@ -244,7 +244,7 @@ impl Symbolizer {
         let sym = Sym {
             name: self.maybe_demangle(name.unwrap_or(sym_name), lang),
             addr: sym_addr,
-            offset: addr - sym_addr,
+            offset: (addr - sym_addr) as usize,
             size: sym_size,
             code_info,
             inlined: inlined.into_boxed_slice(),
@@ -593,7 +593,7 @@ mod tests {
             .unwrap();
 
         let elf_mmap = mmap
-            .constrain(so.data_offset..so.data_offset + so.data.len())
+            .constrain(so.data_offset..so.data_offset + so.data.len() as u64)
             .unwrap();
 
         // Look up the address of the `the_answer` function inside of the shared
@@ -608,7 +608,7 @@ mod tests {
         assert_eq!(syms.len(), 1);
         let sym = syms.first().unwrap();
 
-        let the_answer_addr = unsafe { elf_mmap.as_ptr().add(sym.addr) };
+        let the_answer_addr = unsafe { elf_mmap.as_ptr().add(sym.addr as usize) };
         // Now just double check that everything worked out and the function
         // is actually where it was meant to be.
         let the_answer_fn =
